requirements.txt
hyperliquid-python-sdk>=0.22.0
discord.py>=2.3.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
anthropic>=0.40.0
